+++ src/rtcModule/rtcStats.h
@@ -99,7 +99,7 @@ protected:
             mBwInfo = aBwInfo;
             mBwInfo->bs = 0;
         }
-        void calculate(long periodMs, long newTotalBytes);
+        void calculate(uint64_t periodMs, uint64_t newTotalBytes);
     };
 
     int mScanPeriod;
@@ -117,7 +117,7 @@ protected:
     BwCalculator mConnTxBwCalc;
     void addSample();
     void resetBwCalculators();
-    long long getLongValue(webrtc::StatsReport::StatsValueName name, const webrtc::StatsReport* item);
+    int64_t getLongValue(webrtc::StatsReport::StatsValueName name, const webrtc::StatsReport* item);
     std::string getStringValue(webrtc::StatsReport::StatsValueName name, const webrtc::StatsReport* item);
 public:
     Session& mSession;
